Complete DebugAdapter rename and address CodeRabbit review
The previous commit only recorded the file moves, not the namespace/reference
edits, so HEAD still had 'namespace ClassicAssist.Debug' and the test-project
collision. This commits the actual rename content plus review fixes:

- EvalCondition/InterpolateLogMessage now load globals + locals (shared
  PopulateScope helper) so conditional breakpoints and logpoints can reference
  module globals, imported commands and args.
- Drop the advertised 'uncaught' exception filter that couldn't be honoured
  (IronPython's trace 'exception' event can't distinguish caught/uncaught).
- DapServer.Initialize tears down partial state if TcpListener.Start() fails.
- Validate the debug adapter port (1-65535) before starting; disable the port
  field while the adapter is running so it can't drift from the listener.
- ToggleDebugAdapter catch calls DapServer.Shutdown() before reverting.
